Modified several files and renamed several files. Progress of deserialize has increased...
