--- Unbored/Unbored/Models/ITask.cs
@@ -1,7 +1,0 @@
-namespace Unbored.Models
-{
-    public class ITask
-    {
-        public IList<TaskData>? TaskD { get; set; }
-    }
-}
